requirements.txt
python-dotenv==1.0.1
langchain==0.3.18
langchain_core==0.3.47
langchain_openai==0.3.9
weaviate==0.1.2
pydantic==2.10.6